design: 1차 css 작업 및 파일 생성
