+++ src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Header from "./components/common/Header";
 import Home from "./pages/Home";
 import GlobalStyle from "./styles/globalStyle";
-import SideBar from "./components/common/SideBar";
+import SideBar from "./components/SideBar/SideBar";
 
 const App = () => {
   return (
