+++ src/components/SideBar/OfficeLife.tsx
@@ -1,0 +1,14 @@
+import { useState } from "react";
+import { Routes, Route, useNavigate } from "react-router-dom";
+import companyBlue from "../../assets/icons/sidebar_icon/company_blue_icon.png";
+import bottomBlue from "../../assets/icons/sidebar_icon/bottom_blue_icon.png";
+
+const OfficeLife = () => {
+  return (
+    <div>
+      <span>OfficeLife</span>
+    </div>
+  );
+};
+
+export default OfficeLife;
