+++ src/components/SideBar/Onboarding.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Onboarding = () => {
+  return <div>Onboarding</div>;
+};
+
+export default Onboarding;
