+++ src/components/SideBar/Project.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Project = () => {
+  return <div>Project</div>;
+};
+
+export default Project;
